Move the re-usable pieces of the CPUBackend into a dedicated LLVMIRCodeGen library
- Move the re-usable pieces of the CPUBackend into a dedicated `LLVMIRCodeGen` library

- Introduce `LLVMBackend`, `LLVMCompiledFunction` and `LLVMIRGen` types that can be used as superclasses for different LLVM-based backends

  Concrete backends can derive from those classes and override specific parts of their logic to implement target-specific code generation.

- Each LLVM-based backend can have its own backend-specific libjit now

  The backend should just override `getLibjitBitcode` to achieve that.

- Re-factor the `CPUBackend` to use this new `LLVMIRCodeGen` library
