+++ include/glow/LLVMIRCodeGen/GlowJIT.h
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifndef GLOW_BACKENDS_CPU_GLOWJIT_H
-#define GLOW_BACKENDS_CPU_GLOWJIT_H
+#ifndef GLOW_LLVMIRCODEGEN_GLOWJIT_H
+#define GLOW_LLVMIRCODEGEN_GLOWJIT_H
 
 #include "llvm/ADT/STLExtras.h"
 #include "llvm/ExecutionEngine/ExecutionEngine.h"
@@ -78,4 +78,4 @@ public:
 } // end namespace orc
 } // end namespace llvm
 
-#endif // GLOW_BACKENDS_CPU_GLOWJIT_H
+#endif // GLOW_LLVMIRCODEGEN_GLOWJIT_H
