+++ lib/CMakeLists.txt
@@ -8,6 +8,9 @@ add_subdirectory(ExecutionEngine)
 add_subdirectory(Graph)
 add_subdirectory(IR)
 add_subdirectory(Importer)
+if(GLOW_WITH_LLVMIRCODEGEN)
+    add_subdirectory(LLVMIRCodeGen)
+endif()
 add_subdirectory(Optimizer)
 add_subdirectory(Quantization)
 add_subdirectory(Runtime)
