+++ lib/LLVMIRCodeGen/AllocationsInfo.h
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifndef GLOW_BACKENDS_CPU_ALLOCATIONSINFO_H
-#define GLOW_BACKENDS_CPU_ALLOCATIONSINFO_H
+#ifndef GLOW_LLVMIRCODEGEN_ALLOCATIONSINFO_H
+#define GLOW_LLVMIRCODEGEN_ALLOCATIONSINFO_H
 
 #include "glow/Graph/Nodes.h"
 #include "llvm/IR/Module.h"
@@ -74,4 +74,4 @@ struct AllocationsInfo {
 };
 
 } // namespace glow
-#endif // GLOW_BACKENDS_CPU_ALLOCATIONSINFO_H
+#endif // GLOW_LLVMIRCODEGEN_ALLOCATIONSINFO_H
